feat:convert into function component Search
