--- src/components/Preloader.tsx
@@ -1,13 +1,0 @@
-import React from 'react';
-
-class Preloader extends React.Component {
-  render() {
-    return (
-      <div className="loader-content">
-        <span className="loader"></span>
-      </div>
-    );
-  }
-}
-
-export default Preloader;
