+++ src/components/Search/Search.module.css
@@ -1,0 +1,26 @@
+.search_content {
+  display: flex;
+  justify-content: space-between;
+  padding-top: 20px;
+  padding-bottom: 20px;
+}
+
+.search_content input {
+  width: 100%;
+  border-radius: 10px;
+  margin-right: 20px;
+  border-color: #4a148c;
+  padding: 15px;
+  font-size: 16px;
+}
+
+.search_button {
+  background-color: #4a148c;
+  border: #4a148c;
+  color: #fff;
+  min-width: 200px;
+}
+
+.search_button:hover {
+  background-color: #7b1fa2;
+}
