+++ src/index.css
@@ -101,24 +101,6 @@ button:focus-visible {
   display: block;
 }
 
-.loader-content {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 50vh;
-}
-
-.loader {
-  width: 48px;
-  height: 48px;
-  border: 5px dotted #fff;
-  border-radius: 50%;
-  display: inline-block;
-  position: relative;
-  box-sizing: border-box;
-  animation: rotation 2s linear infinite;
-}
-
 @keyframes rotation {
   0% {
     transform: rotate(0deg);
